tools/translate/requirements.txt
PyPDF2
pillow
pytesseract
pdf2image
argparse